Minor typoreadagrams.
Right at the bottom, it speaks of "making changes click on the basic setup". Was this meant to say "making changes to the basic setup"?
Also, there's reference to "DCC-EX Labs". Is this an official or preferred way to reference the group? On the homepage it says "DCC++ EX" is the organisation doing this project.